first attempt at rewrite of gui
