+++ .gitignore
@@ -273,3 +273,4 @@ __pycache__/
 
 # pip nonsense
 src/*
+circleguard/gui_v2.py
